separate constraints
