+++ constraints_u.py
@@ -1,0 +1,41 @@
+#
+# The MIT License
+#
+# Copyright (c) 2022 Interstellar Technologies Inc.
+#
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files
+# (the "Software"), to deal in the Software without restriction,
+# including without limitation the rights to use, copy, modify, merge,
+# publish, distribute, sublicense, and/or sell copies of the Software,
+# and to permit persons to whom the Software is furnished to do so,
+# subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+#
+
+# constraints_u.py
+# constraints about user conditions
+
+from user_constraints import equality_user, inequality_user
+from utils import jac_fd
+
+def equality_jac_user(xdict, pdict, unitdict, condition):
+    """Jacobian of user-defined equality constraint."""
+    if equality_user(xdict, pdict, unitdict, condition) is not None:
+        return jac_fd(equality_user, xdict, pdict, unitdict, condition)
+
+
+def inequality_jac_user(xdict, pdict, unitdict, condition):
+    """Jacobian of user-defined inequality constraint."""
+    if inequality_user(xdict, pdict, unitdict, condition) is not None:
+        return jac_fd(inequality_user, xdict, pdict, unitdict, condition)
